Refactoring with testInitializeAudit and testFederalAwards
