+++ backend/cypress/e2e/secondary-auditors.cy.js
@@ -1,50 +1,20 @@
 import 'cypress-file-upload';
-import { testValidAccess } from '../support/check-access.js';
-import { testValidEligibility } from '../support/check-eligibility.js';
-import { testValidAuditeeInfo } from '../support/auditee-info.js';
-import { testValidGeneralInfo } from '../support/general-info.js';
-import { testReportIdNotFound } from '../support/dissemination-table.js';
+import { testFederalAwards } from '../support/federal-awards.js';
 import { testFileUploadMsg } from '../support/file-uploaded-msg.js';
+import { testLoginGovLogin } from '../support/login-gov.js';
 import {
     testWorkbookSecondaryAuditors,
-  testWorkbookFederalAwards,
 } from '../support/workbook-uploads.js';
 
 describe('Secondary Auditors page', () => {
   before(() => {
-    cy.session('login-session', () => {
-      cy.visit('/');
-      cy.login();
-    });
+    cy.visit('/');
+    cy.url().should('include', '/');
+    testLoginGovLogin();
   });
 
   it('Secondary auditors uploads successfully', () => {
-    cy.visit('/');
-
-    cy.url().should('include', '/');
-
-    cy.get('label[for=check-start-new-submission]').click();
-
-    cy.get('.usa-button').contains('Accept and start').click();
-
-    cy.url().should('match', /\/report_submission\/eligibility\/$/);
-
-    testValidEligibility();
-
-    testValidAuditeeInfo();
-
-    testValidAccess();
-
-    // Report should not yet be in the dissemination table
-    cy.url().then(url => {
-      const reportId = url.split('/').pop();
-      testReportIdNotFound(reportId);
-    });
-
-    testValidGeneralInfo();
-
-    cy.get(".usa-link").contains("Federal Awards").click();
-    testWorkbookFederalAwards(false);
+    testFederalAwards();
 
     cy.get(".usa-link").contains("Secondary Auditors").click();
     testWorkbookSecondaryAuditors(false);
@@ -54,4 +24,3 @@ describe('Secondary Auditors page', () => {
     testFileUploadMsg('Edit the Secondary Auditors');
   });
 });
-
