+++ backend/cypress/support/full-submission.js
@@ -1,17 +1,13 @@
 import { testCrossValidation } from './cross-validation.js';
 import { testLoginGovLogin } from './login-gov.js';
 import { testLogoutGov } from './logout-gov.js';
-import { testValidAccess } from './check-access.js';
-import { testValidEligibility } from './check-eligibility.js';
-import { testValidAuditeeInfo } from './auditee-info.js';
-import { testValidGeneralInfo } from './general-info.js';
 import { testAuditInformationForm } from './audit-info-form.js';
 import { testPdfAuditReport } from './report-pdf.js';
 import { testAuditorCertification } from './auditor-certification.js';
 import { testAuditeeCertification } from './auditee-certification.js';
 import { testReportIdFound, testReportIdNotFound } from './dissemination-table.js';
 import { testTribalAuditPublic, testTribalAuditPrivate } from './tribal-audit-form.js';
-
+import { testInitializeAudit } from './initialize-audit.js';
 import {
   testWorkbookFederalAwards,
   testWorkbookNotesToSEFA,
@@ -34,32 +30,7 @@ export function testFullSubmission(isTribal, isPublic) {
   // Logs in with Login.gov'
   testLoginGovLogin();
 
-  // Moves on to the eligibility screen
-  // check the terms and conditions link and click "Accept and start..."
-  //
-  // this click actually goes to the "terms and conditions" link which
-  // brings up a modal
-  cy.get('label[for=check-start-new-submission]').click();
-  cy.get('.usa-button').contains('Accept and start').click();
-  cy.url().should('match', /\/report_submission\/eligibility\/$/);
-
-  // Completes the eligibility screen
-  testValidEligibility(isTribal);
-
-  // Now the auditee info screen
-  testValidAuditeeInfo();
-
-  // Now the accessandsubmission screen
-  testValidAccess();
-
-  // Report should not yet be in the dissemination table
-  cy.url().then(url => {
-    const reportId = url.split('/').pop();
-    testReportIdNotFound(reportId);
-  });
-
-  // Fill out the general info form
-  testValidGeneralInfo();
+  testInitializeAudit(isTribal);
 
   // Upload all the workbooks. Don't intercept the uploads, which means a file will make it into the DB.
   cy.get(".usa-link").contains("Federal Awards").click();
@@ -155,7 +126,7 @@ export function testFullSubmission(isTribal, isPublic) {
     // The report ID should be found in the Completed Audits table
     cy.get('.usa-table').contains(
       'caption',
-      'The audits listed below have been submitted to the FAC for processing and may not be edited.',
+      /audits are complete/
     ).siblings().contains('td', reportId);
 
     // The Report should not be in the dissemination table
