+++ backend/cypress/support/initialize-audit.js
@@ -1,0 +1,34 @@
+import { testValidAccess } from './check-access.js';
+import { testValidEligibility } from './check-eligibility.js';
+import { testValidAuditeeInfo } from './auditee-info.js';
+import { testValidGeneralInfo } from './general-info.js';
+import { testReportIdNotFound } from './dissemination-table.js';
+
+export function testInitializeAudit(isTribal=false) {
+  // Moves on to the eligibility screen
+  // check the terms and conditions link and click "Accept and start..."
+  //
+  // this click actually goes to the "terms and conditions" link which
+  // brings up a modal
+  cy.get('label[for=check-start-new-submission]').click();
+  cy.get('.usa-button').contains('Accept and start').click();
+  cy.url().should('match', /\/report_submission\/eligibility\/$/);
+
+  // Completes the eligibility screen
+  testValidEligibility(isTribal);
+
+  // Now the auditee info screen
+  testValidAuditeeInfo();
+
+  // Now the accessandsubmission screen
+  testValidAccess();
+
+  // Report should not yet be in the dissemination table
+  cy.url().then(url => {
+    const reportId = url.split('/').pop();
+    testReportIdNotFound(reportId);
+  });
+
+  // Fill out the general info form
+  testValidGeneralInfo();
+}
